+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.6.2] - 2022-11-03
 ### Added
 - Override update notifications preferences [#270](https://github.com/rokwire/groups-building-block/issues/270)
 - Group notification internal API [#241](https://github.com/rokwire/groups-building-block/issues/241)